+++ lactosafe/lib/src/controller/Info_food_controllers/risk_measuring_emoji_controller.dart
@@ -1,6 +1,6 @@
 import 'package:LactoSafe/src/shared/app_images.dart';
 
-String riskMeasuringEmoji({required double foodRisk}) {
+String riskMeasuringEmoji({required int foodRisk}) {
   switch (foodRisk) {
     case 0:
       return AppImages.smileEmoji;
